requirements.txt
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
python-dotenv==1.0.0
requests==2.28.2
soupsieve==2.4
urllib3==1.26.15
selenium~=4.8.2